[android] gridview layout setup encapsulated
Number of columns available in vertical mode is always 3.
Number of columns available in horizontal mode depends on the size of the screen, currently the same number of inches available on the screen diagonal, will be fun to test on a tv to see how many pictures we get, might have to add an extra case for more than N inches.
layout files renamed